Student details app (flask_jwt_extended)

+++ StudentDetails_SQLAlchemy/resources/student.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity, fresh_jwt_required
 from models.student import StudentModel
 
 class Student(Resource):
@@ -7,13 +7,14 @@ class Student(Resource):
     parser.add_argument("cgpa", required=True, type=float, help="This filed cant be left empty")
     parser.add_argument("department_id", required=True, type=int, help="Student's registered department")
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         student = StudentModel.find_by_name(name)
         if student:
             return {"Student Details": student.json()}
         return {"message":"Student not found"},404
 
+    @fresh_jwt_required
     def post(self, name):
 
         student = StudentModel.find_by_name(name)
@@ -28,7 +29,11 @@ class Student(Resource):
             return {"message":"An error occurred while inserting the details."}, 500
         return {"message":"Student details added successfully","Student Details":student.json()}, 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {"message": "Admin privileges required "}, 401
         student = StudentModel.find_by_name(name)
         if student:
             student.delete_from_db()
@@ -50,5 +55,13 @@ class Student(Resource):
 
 
 class StudentList(Resource):
+    @jwt_optional
     def get(self):
-        return {"Students": [student.json() for student in StudentModel.query.all()]}
+        user_id = get_jwt_identity()
+        student_list = [student.json() for student in StudentModel.find_all()]
+        if user_id:
+            return {"Students": student_list}, 200
+
+        return {"names": [student['name'] for student in student_list],
+                "message": "Please log in for more information"}, 200
+
